requirements.txt
django
djangorestframework
python-decouple
pillow
psycopg2-binary
